+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Added
 - Flow ``role_update_irods_batch`` for updating user roles in iRODS (#60)
 - Inherited category owners into iRODS user groups in ``project_create`` (#59)
 - Bulk updating of user roles in ``project_update`` (#61)
+- iRODS task ``RemoveDataObjectTask``
 
 Changed
 -------